Fix handling of NOW() for detached sessions
Summary: The handling of NOW() uses a cached time value that should have been set at the start of the statement.  This value wasn't being set for the detached session so all subsequent statements continued to use the same time.  Additionally, if the user set the current timestamp (SET TIMESTAMP = <value>) that NOW() should use, it was being ignored.  Also, the processlist was using the connection thread's start time instead of the srv_session's start time.

Reviewed By: abal147

Differential Revision: D14313267

fbshipit-source-id: a03502b4501