Allow KILL ID where ID is a connection with an attached srv_session
Summary: We originally did not allow KILLs on a connection id with an attached srv_session because we felt it was safer to force KILLing the srv_session.  However in discussions we decided that any problem would be very rare and it would be more confusing to require the caller to specify the srv_session id.  In addition we can now correctly kill conflicting connections

Reviewed By: abal147

Differential Revision: D14313443

fbshipit-source-id: 77e4851caf8