Remove the RPC_ID from the info portion of the ok packet
Summary: At one point we returned the rpc_id in the info portion of the info packet.  Once we switched to using response attributes we needed to leave it in until MyRouter looked for the data in the response attributes.  But then it was forgotten.  Now when turning on rpc_protocol, it causes a problem in the main.alter_table test as that test turns on the display of the info field.  Remove the rpc_id from the info portion.

Reviewed By: abal147

Differential Revision: D14313578

fbshipit-source-id: 2bbd6e57b7b