mysqlslap honors protocol
Summary: I find out in my testing that mysqlslap doesn't really use the --protocol parameter in its workers. The bigger problem seems to be that it doesn't apply some of its parameters in the worker thread, --protocol is just one of them. I'm not sure I have time to test all the parameters here so I'm just adding the fix for --protocol. The testing for mysqlslap seems be rather minimal in MySQL in general.

Reviewed By: hermanlee

Differential Revision: D15010933

fbshipit-source-id: d05be7bdf55